Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #7:

This patch reduces the compiler warning count on Win64 by an additional
200, as reported by VS2013.

+++ sql-common/sql_string.cc
@@ -138,7 +138,8 @@ bool String::set_real(double num,uint decimals, const CHARSET_INFO *cs)
   str_charset=cs;
   if (decimals >= NOT_FIXED_DEC)
   {
-    len= my_gcvt(num, MY_GCVT_ARG_DOUBLE, sizeof(buff) - 1, buff, NULL);
+    len= my_gcvt(num, MY_GCVT_ARG_DOUBLE, static_cast<int>(sizeof(buff)) - 1,
+                 buff, NULL);
     return copy(buff, len, &my_charset_latin1, cs, &dummy_errors);
   }
   len= my_fcvt(num, decimals, buff, NULL);
